tests: Move run_swf to util/runner
